requirements.txt
# 测试框架核心依赖
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==12.0
pytest-timeout==2.2.0

# Robot Framework
robotframework==6.1.1
robotframework-seleniumlibrary==6.8.0
robotframework-requests==0.9.7

# Selenium
selenium==4.15.2
webdriver-manager==4.0.2

# Playwright
playwright==1.40.0
robotframework-browser==18.0.0

# API 测试
requests==2.31.0

# 其他工具
python-dotenv==1.0.0
requests==2.32.3
# allure-pytest 已移除（与 Python 3.11 兼容性问题），使用 pytest-html 代替


